requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
langchain-community
langchain-huggingface
langchain-pinecone
flask
pypdf
python-dotenv

# -e .